Added a few more searches of popular tablets based upon browsing hashtags.
I think later this semester I may be able to remove #surface or #surface2, depending on which one gets dropped (if one does) after the second surfaces release. As of the moment I'm not too sure of a good way of to look through android tablets since its such a spread market, and even though the nexus 10 is google's next version it seems like its not getting a lot of hype. Thankfully ipad is #ipad and gets a lot of data.