get dataframe of top N source
This implementation is about getting a dataframe of top N source.
For me, this implementation is little bit confused due to various type of
source such as twitterfeed, dlvr.it, and IFTTT. What are those?
Originally, I just want to know what kind of device do people who was
twitting this tweet about devices use.
I think I need to know what those are such as dlvr.it and how can treat
those kind of things? Is it based on which device?